WORKING PAGES!!!
this update adds working pages to the program!
we have shuffled widgets around to account for the pages changing.
last change that was made is that we are aware that the clear button does not work on other pages yet. we are working on fixing that.
TTMv2 will mark a new branch of the program as much has changed.
ttmv1 will stay the same from this point onwards.